database-service/requirements.txt
flask==3.0.2
pymongo==4.6.1
redis==5.0.1
python-dotenv==1.0.1 